docs: transfer the rouletteNumber constant to a separate module
