+++ roulette/models/multiCompleteModel.js
@@ -1,8 +1,8 @@
 // import "./menuModel.js";
 // import { minBet } from "./menuModel.js";
 // import { maxBet } from "./menuModel.js";
-// import { payoutRatios } from "./menuModel.js";
-// import { rouletteNumber } from "./menuModel.js";
+// import { payoutRatios } from "./consts/payoutRatios.js";
+// import { rouletteNumber } from "./consts/rouletteNumber.js";
 import { chipsNeededForABet } from "./menuModel.js";
 import { countPositions } from "./menuModel.js";
 import { oddDozColum } from "./menuModel.js";
