+++ roulette/models/trackModel.js
@@ -3,7 +3,7 @@
 // import { maxBet } from "./menuModel.js";
 import { rouletteSeries } from "./menuModel.js";
 import { rltPos } from "./menuModel.js";
-import { payoutRatios } from "./menuModel.js";
+import { payoutRatios } from "./consts/payoutRatios.js";
 // import { seriesCalc } from "./menuModel.js";
 // import { calculateBTN } from "./menuModel.js";
