+++ roulette/views/menu/menuView.html
@@ -62,6 +62,7 @@
 
     <script src="../../models/consts/min_max.js" type="module"></script>
     <script src="../../models/consts/payoutRatios.js" type="module"></script>
+    <script src="../../models/consts/rouletteNumber.js" type="module"></script>
     <script src="../../models/menuModel.js" type="module"></script>
 </body>
